requirements.txt
asgiref             == 3.11.1
Django              == 5.2.11
djangorestframework == 3.16.1
pip                 == 24.0
setuptools          == 65.5.0
sqlparse            == 0.5.5
tzdata              == 2025.3
certifi             == 2026.1.4
charset-normalizer  == 3.4.4
django-redis        == 6.0.0
idna                == 3.11
polyline            == 2.0.4
redis               == 7.2.0
requests            == 2.32.5
urllib3             == 2.6.3
psycopg2-binary     == 2.9.11
httpx               == 0.28.1
pytest              == 9.0.2
pytest-django       == 4.12.0
pytest-mock         == 3.15.1
factory_boy         == 3.3.3